+++ collectors/nvd/tests/test_collectors.py
@@ -3,7 +3,7 @@ from django.utils import timezone
 
 from collectors.framework.models import CollectorMetadata
 from collectors.nvd.collectors import NVDCollector
-from osidb.models import Flaw, FlawCVSS, Snippet
+from osidb.models import Flaw, FlawCVSS, FlawReference, Snippet
 from osidb.tests.factories import FlawCVSSFactory, FlawFactory
 
 pytestmark = pytest.mark.integration
@@ -235,64 +235,76 @@ class TestNVDCollector:
         cve = "CVE-2017-7542"
 
         snippet_content = {
-            "cve_ids": ["CVE-2017-7542"],
-            "cvss2": {
-                "score": 4.9,
-                "issuer": "nvd@nist.gov",
-                "vector": "AV:L/AC:L/Au:N/C:N/I:N/A:C",
-            },
-            "cvss3": {
-                "score": 5.5,
-                "issuer": "nvd@nist.gov",
-                "vector": "CVSS:3.0/AV:L/AC:L/PR:L/UI:N/S:U/C:N/I:N/A:H",
-            },
-            "cwe_id": "CWE-190|CWE-835",
+            "cve_id": cve,
+            "cvss_scores": [
+                {
+                    "score": 4.9,
+                    "issuer": FlawCVSS.CVSSIssuer.NIST,
+                    "vector": "AV:L/AC:L/Au:N/C:N/I:N/A:C",
+                    "version": FlawCVSS.CVSSVersion.VERSION2,
+                },
+                {
+                    "score": 5.5,
+                    "issuer": FlawCVSS.CVSSIssuer.NIST,
+                    "vector": "CVSS:3.0/AV:L/AC:L/PR:L/UI:N/S:U/C:N/I:N/A:H",
+                    "version": FlawCVSS.CVSSVersion.VERSION3,
+                },
+            ],
+            "cwe_id": "(CWE-190|CWE-835)",
             "description": "The ip6_find_1stfragopt function in net/ipv6/output_core.c in the Linux kernel through 4.12.3 allows local users to cause a denial of service (integer overflow and infinite loop) by leveraging the ability to open a raw socket.",
             "references": [
                 {
                     "url": "https://nvd.nist.gov/vuln/detail/CVE-2017-7542",
-                    "type": "SOURCE",
+                    "type": FlawReference.FlawReferenceType.SOURCE,
                 },
                 {
                     "url": "http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=6399f1fae4ec29fab5ec76070435555e256ca3a6",
-                    "type": "EXTERNAL",
+                    "type": FlawReference.FlawReferenceType.EXTERNAL,
                 },
                 {
                     "url": "http://www.debian.org/security/2017/dsa-3927",
-                    "type": "EXTERNAL",
+                    "type": FlawReference.FlawReferenceType.EXTERNAL,
                 },
                 {
                     "url": "http://www.debian.org/security/2017/dsa-3945",
-                    "type": "EXTERNAL",
+                    "type": FlawReference.FlawReferenceType.EXTERNAL,
                 },
                 {"url": "http://www.securityfocus.com/bid/99953", "type": "EXTERNAL"},
                 {
                     "url": "https://access.redhat.com/errata/RHSA-2017:2918",
-                    "type": "EXTERNAL",
+                    "type": FlawReference.FlawReferenceType.EXTERNAL,
                 },
                 {
                     "url": "https://access.redhat.com/errata/RHSA-2017:2930",
-                    "type": "EXTERNAL",
+                    "type": FlawReference.FlawReferenceType.EXTERNAL,
                 },
                 {
                     "url": "https://access.redhat.com/errata/RHSA-2017:2931",
-                    "type": "EXTERNAL",
+                    "type": FlawReference.FlawReferenceType.EXTERNAL,
                 },
                 {
                     "url": "https://access.redhat.com/errata/RHSA-2018:0169",
-                    "type": "EXTERNAL",
+                    "type": FlawReference.FlawReferenceType.EXTERNAL,
                 },
                 {
                     "url": "https://github.com/torvalds/linux/commit/6399f1fae4ec29fab5ec76070435555e256ca3a6",
-                    "type": "EXTERNAL",
+                    "type": FlawReference.FlawReferenceType.EXTERNAL,
                 },
                 {
                     "url": "https://help.ecostruxureit.com/display/public/UADCE725/Security+fixes+in+StruxureWare+Data+Center+Expert+v7.6.0",
-                    "type": "EXTERNAL",
+                    "type": FlawReference.FlawReferenceType.EXTERNAL,
                 },
-                {"url": "https://usn.ubuntu.com/3583-1/", "type": "EXTERNAL"},
-                {"url": "https://usn.ubuntu.com/3583-2/", "type": "EXTERNAL"},
+                {
+                    "url": "https://usn.ubuntu.com/3583-1/",
+                    "type": FlawReference.FlawReferenceType.EXTERNAL,
+                },
+                {
+                    "url": "https://usn.ubuntu.com/3583-2/",
+                    "type": FlawReference.FlawReferenceType.EXTERNAL,
+                },
             ],
+            "source": Snippet.Source.NVD,
+            "title": "placeholder only, see description",
         }
 
         # Default data
@@ -306,12 +318,12 @@ class TestNVDCollector:
         nvdc.collect(cve)
 
         all_snippets = Snippet.objects.filter(
-            source=Snippet.Source.NVD, content__cve_ids=[cve]
+            source=Snippet.Source.NVD, content__cve_id=cve
         )
         snippet = all_snippets.first()
 
         assert len(all_snippets) == 1
-        assert len(snippet.content) == 6
+        assert len(snippet.content) == 7
 
         for key, value in snippet_content.items():
             assert snippet.content[key] == value